server/requirements.txt
flask==3.0.0
pyserial==3.5
python-dotenv==1.0.0
requests==2.31.0
sounddevice==0.4.6
scipy==1.11.4
keyboard==0.13.5
